requirements.txt
streamlit
Pillow
requests
openai
flask
python-dotenv
